:recycle: Code review move join_node from internal/models to internal/cluster
