+++ internal/cluster/main.go
@@ -7,7 +7,6 @@ import (
 
 	"github.com/vladopajic/go-actor/actor"
 
-	"link-society.com/flowg/internal/models"
 	"link-society.com/flowg/internal/utils/proctree"
 
 	"github.com/hashicorp/memberlist"
@@ -17,7 +16,7 @@ type ManagerOptions struct {
 	NodeID string
 	Cookie string
 
-	ClusterJoinNode *models.ClusterJoinNode
+	ClusterJoinNode *ClusterJoinNode
 
 	AutomaticClusterFormation bool
